Updated jest configuration to add jest-html-reporter. Installed jest-html-reporter package. Developed delete method of JobConfigCommand to parse the job ids and verify that they are valid ints and if not throw an error. Updated GeneratedBearerTokenRESTAuthHandler test class to improve test descriptions. Updated ConfigValidator test class to improve test descriptions. Updated RESTLoaderJobConfigValidator test class to improve test descriptions. Updated HashiCorpVault test class to improve test descriptions. Added test to HashiCorpVault test class to verify failure when an invalid key is used to read from vault. Updated LoaderJob test class to improve test descriptions. Updated RESTLoaderJob test class to improve test descriptions. Added tests to RESTLoaderJob to verify correct operation when using GET method. Updated Scheduler test class to improve test descriptions. Updated BufferedLogApi test class to improve test descriptions. Added test to BufferedLogApi test class to verify that it can be correctly stopped. Updated FileQueue test class to improve test descriptions. Updated JobConfigApi test class to improve test descriptions. Updated LogApi  test class to improve test descriptions. Updated config test class to improve test descriptions. Updated logger test class to improve test descriptions. Developed IngestionCommand test class. Developed JobConfigCommand test class.

+++ src/command/JobConfigCommand.js
@@ -75,7 +75,12 @@ class JobConfigCommand extends Command {
       all
     } = argv
     if (ids.length) {
-      for (const id of ids) {
+      console.log('ids: ', ids)
+      for (let id of ids) {
+        id = parseInt(id)
+        if (isNaN(id)) {
+          throw new Error('Please enter valid uint64 numbers')
+        }
         console.log(`Deleting job with id: ${id}...`)
         await this.jobConfig.jobConfigApi.delete(id)
       }
